removing DebugMode class from sidescroller
